Scanning hardware interface separation, still a few things to iron out.
